[fix] move StoreKey constant to types file so it can be exported and used by other implementations of the DVC Cache Store
